backend/source/requirements.txt
Django
environs
psycopg2
pytest
pytest-django
pylint-django
djangorestframework
markdown      
django-filter 
django-cors-headers
boto3
django-storages
gunicorn